+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!--## [Unreleased]-->
+## [4.3.0] - 2019-09-13
 ### Added
 - includeCompleted flag for v1/characters/{character_id}/industry/jobs.
 - /v2/universe/stations/{station_id}}/ URL and typings.
@@ -91,7 +92,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial codebase, most copied from <https://github.com/Ionaru/evie>
 - Setup for this project
 
-[Unreleased]: https://github.com/Ionaru/eve-utils/compare/4.2.0...HEAD
+[Unreleased]: https://github.com/Ionaru/eve-utils/compare/4.3.0...HEAD
+[4.3.0]: https://github.com/Ionaru/eve-utils/compare/4.2.0...4.3.0
 [4.2.0]: https://github.com/Ionaru/eve-utils/compare/4.1.3...4.2.0
 [4.1.3]: https://github.com/Ionaru/eve-utils/compare/4.1.2...4.1.3
 [4.1.2]: https://github.com/Ionaru/eve-utils/compare/4.1.1...4.1.2